Split OandaBrokerage as partial class
