+++ Brokerages/QuantConnect.Brokerages.csproj
@@ -199,6 +199,7 @@
     <Compile Include="Oanda\Framework\Common.cs" />
     <Compile Include="Oanda\Framework\EndpointResolver.cs" />
     <Compile Include="Oanda\OandaBrokerage.cs" />
+    <Compile Include="Oanda\OandaBrokerage.Rest.cs" />
     <Compile Include="Oanda\OandaBrokerageFactory.cs" />
     <Compile Include="Oanda\OandaSymbolMapper.cs" />
     <Compile Include="Oanda\Server.cs" />
